% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margplot.R
\name{margplot}
\alias{margplot}
\title{Plotting marginal effects for interactions in nonlinear probability and count models}
\usage{
margplot(
  model,
  vars,
  data,
  hyps = "means",
  foc,
  mod,
  modlevels,
  modnames = NULL,
  sm = T
)
}
\arguments{
\item{model}{The estmiated model object.
Currently, this may take logit or Poisson model objects
estimated using the 'stats' package or negative binomial
model objects tested using the 'MASS" package.}

\item{vars}{The interacting variable names.}

\item{data}{The data frame on which the estimated model is based.}

\item{hyps}{User-specified levels of the predictor variables
for displaying marginal effects. By default,
this is specified at the mean values of all included covariates.}

\item{foc}{Name of the variable that will be the focal (i.e., x-axis) variable}

\item{mod}{Name of hte variable that will be the moderator (i.e., legend) variable}

\item{modlevels}{A vector of hypothetical levels of the moderator
at which to display marginal effects.}

\item{modnames}{If the user wants to call the moderator levels something other
Than what is provided in the data itself, the user can supply a vector of names here.}

\item{sm}{Stands for "small multiple". If TRUE, marginal effects will be displayed
as small multiple (i.e., trellis) plots. if FALSE, all marginal effects will appear
on a single plot.}
}
\value{

}
\description{
Plotting marginal effects for interactions in nonlinear probability and count models
}
\examples{
#Building from the modglm example, this outputs a ggplot object plotting the effect of x1 on
#the count of y across males and females:
margplot(model=pois, vars=c("x1","female"),foc="x1", mod="female", modlevels=c(0,1),data=df, hyps=c(1,0,-2,0,0),sm=F,modnames=c("Male","Female"))
}
