% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modglm.R
\name{modglm}
\alias{modglm}
\title{Computing Interactions in nonlinear probability and count models}
\usage{
modglm(model, vars, data, hyps = "means", plotby = NULL, type = "cpd")
}
\arguments{
\item{model}{The estmiated model object. Currently, this may ntake logit or Poisson model objects estimated using the 'stats' package or negative binomial mdoel objects ested using the 'MASS" package.}

\item{vars}{The interacting variable names.}

\item{data}{The data frame on which the estimated model is based.}

\item{hyps}{User-specified levels of the predictor variables for evaluating the interaction function. By default, this is specified at the mean values of all included covariates.}

\item{plotby}{An option to view the interaction effect estimates plot across categories of another variable. Default is NULL.}

\item{type}{The type of interaction being estimated. Options are ""cpd", "fd", and "dd". "cpd" indicates cross-partial derivative, and should be used when both varaibles are continuous. "fd" indicates finite difference, and should be used when one variable is continuous and the other is discrete. "dd" indicates double (finite) difference, and should be used if when both variables are categorical.}
}
\value{

}
\description{
Computing Interactions in nonlinear probability and count models
}
